Refactors several of smaller helper services.  Unit test for Request obj.
Also changes the API for the gem a bit.  See updates to the README.

This was an attempt at unit testing our Request class.  It is very
difficult and perhaps that is an issue with design.  Rather than
spending too much time here, I think we can cover this in our
integration tests.  Come back to this test if we find problems.

Another option is digging into Faraday and finding out why the
test stub object is not logging the POST body.  Time permitting :/